Grandes ajustes( Animacao e comeco de refatoracao do codigo)
Adicao de animacao a barra de força da senha
Separacao de arquivos de estilizacao
Remocao de codigos inutilizados
