+++ package.json
@@ -12,11 +12,12 @@
     "@react-native-community/slider": "4.2.4",
     "expo": "~47.0.12",
     "expo-checkbox": "~2.2.2",
+    "expo-clipboard": "~4.0.1",
     "expo-status-bar": "~1.4.2",
     "react": "18.1.0",
     "react-native": "0.70.5",
-    "react-native-vector-icons": "^9.2.0",
-    "expo-clipboard": "~4.0.1"
+    "react-native-reanimated": "~2.12.0",
+    "react-native-vector-icons": "^9.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
